+++ src/main/kotlin/io/github/t45k/kmp_trial/engine/PresentationEngine.kt
@@ -1,0 +1,88 @@
+package io.github.t45k.kmp_trial.engine
+
+import androidx.compose.animation.core.tween
+import androidx.compose.animation.slideInHorizontally
+import androidx.compose.animation.slideOutHorizontally
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.Text
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.compose.ui.window.Window
+import androidx.compose.ui.window.application
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import io.github.t45k.kmp_trial.api.Presentation
+
+fun handlePresentation(presentation: Presentation) = application {
+    val navController = rememberNavController()
+    Window(
+        onCloseRequest = ::exitApplication,
+        title = "Slide.kt",
+        onKeyEvent = moveSlideEvent(navController),
+    ) {
+        NavHost(navController, startDestination = "1") {
+            presentation.slides.forEachIndexed { index, slide ->
+                composable(
+                    route = (index + 1).toString(),
+                    enterTransition = {
+                        slideInHorizontally(
+                            initialOffsetX = { fullWidth -> fullWidth },
+                            animationSpec = tween()
+                        )
+                    },
+                    exitTransition = {
+                        slideOutHorizontally(
+                            targetOffsetX = { fullWidth -> -fullWidth },
+                            animationSpec = tween()
+                        )
+                    },
+                    popEnterTransition = {
+                        slideInHorizontally(
+                            initialOffsetX = { fullWidth -> -fullWidth },
+                            animationSpec = tween()
+                        )
+                    },
+                    popExitTransition = {
+                        slideOutHorizontally(
+                            targetOffsetX = { fullWidth -> fullWidth },
+                            animationSpec = tween()
+                        )
+                    }
+                ) {
+                    Box(
+                        Modifier.fillMaxWidth().padding(16.dp),
+                        contentAlignment = Alignment.Center
+                    ) {
+                        Text(
+                            text = slide.title,
+                            fontSize = 64.sp,
+                            fontWeight = FontWeight.Bold,
+                        )
+                    }
+
+                    Column(
+                        modifier = Modifier.fillMaxSize(),
+                        horizontalAlignment = Alignment.CenterHorizontally,
+                        verticalArrangement = Arrangement.Center,
+                    ) {
+                        slide.paragraphs.forEach { parText ->
+                            Text(
+                                text = parText,
+                                fontSize = 32.sp,
+                            )
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
